backend/requirements-prod.txt
# Additional packages for production deployment
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
